1. 修复vnc协议部分情况下可能异常问题 2. 修复复制、粘贴目录异常问题 3. 打包信息更新 4. 针对文件菜单比较多的问题优化，仅显示可操作菜单 5. 其他细节处理

+++ src/main/java/cn/oyzh/easyshell/fx/file/ShellFileTableView.java
@@ -882,13 +882,6 @@ public abstract class ShellFileTableView<C extends ShellFileClient<E>, E extends
     @Override
     public List<? extends MenuItem> getMenuItems() {
         List<E> files = this.getFilterSelectedItems();
-        // if (CollectionUtil.isEmpty(files)) {
-        //     return Collections.emptyList();
-        // }
-        // // 检查是否包含无效文件
-        // if (this.checkInvalid(files)) {
-        //     return Collections.emptyList();
-        // }
         List<MenuItem> menuItems = new ArrayList<>();
         // 创建文件
         FXMenuItem touchFile = MenuItemHelper.touchFile("12", this::touch);
@@ -936,10 +929,11 @@ public abstract class ShellFileTableView<C extends ShellFileClient<E>, E extends
         refreshFile.setAccelerator(KeyboardUtil.refresh_keyCombination);
         menuItems.add(refreshFile);
         // 删除文件
-        FXMenuItem deleteFile = MenuItemHelper.deleteFile("12", () -> this.deleteFile(files));
-        deleteFile.setAccelerator(KeyboardUtil.delete_keyCombination);
-        deleteFile.setDisable(files.isEmpty());
-        menuItems.add(deleteFile);
+        if (!files.isEmpty()) {
+            FXMenuItem deleteFile = MenuItemHelper.deleteFile("12", () -> this.deleteFile(files));
+            deleteFile.setAccelerator(KeyboardUtil.delete_keyCombination);
+            menuItems.add(deleteFile);
+        }
         return menuItems;
     }
